refactored plugins
