+++ lua/plugins/lualine.lua
@@ -4,4 +4,14 @@ return {
   opts = {
     theme = 'tokyonight',
   },
+  config = function()
+    require('lualine').setup {
+      sections = {
+        lualine_x = {
+          -- Other lualine components in "x" section
+          { require('mcphub.extensions.lualine') },
+        },
+      },
+    }
+  end
 }
